Extract the manual test timer for reuse
The deterministic UserThread timer replacement was nested inside
TxBroadcasterTest and inaccessible to other tests. Move it to a shared
test utility unchanged, with accessors instead of the private field
reads.
